Support themodernproper (#244)

+++ crates/schema-org/src/image_object.rs
@@ -9,7 +9,7 @@ use crate::field::{
     ImageObjectPublisherFieldEnum, ImageObjectSdPublisherFieldEnum, ImageObjectSizeFieldEnum,
     ImageObjectSubjectOfFieldEnum, ImageObjectWidthFieldEnum,
 };
-use crate::helpers::one_or_many;
+use crate::helpers::{one_or_many, one_or_many_string_or_num};
 use crate::{Comment, InteractionCounter, MediaObject, Thing};
 
 ///<https://schema.org/encodingFormat>
@@ -215,9 +215,9 @@ pub struct ImageObject {
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub thumbnail_url: Vec<String>,
     ///<https://schema.org/copyrightYear>
-    #[serde(default, deserialize_with = "one_or_many")]
+    #[serde(default, deserialize_with = "one_or_many_string_or_num")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub copyright_year: Vec<f32>,
+    pub copyright_year: Vec<i32>,
     ///<https://schema.org/citation>
     #[serde(default, deserialize_with = "one_or_many")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
