+++ crates/schema-org/src/video_object.rs
@@ -7,7 +7,7 @@ use crate::field::{
     VideoObjectKeywordsFieldEnum, VideoObjectSdPublisherFieldEnum, VideoObjectSubjectOfFieldEnum,
     VideoObjectVideoFieldEnum,
 };
-use crate::helpers::one_or_many;
+use crate::helpers::{one_or_many, one_or_many_string_or_num};
 use crate::{
     AggregateRating, ImageObject, InteractionCounter, MediaObject, Organization, Person, Review,
     Thing,
@@ -240,9 +240,9 @@ pub struct VideoObject {
     #[serde(skip_serializing_if = "Vec::is_empty")]
     pub thumbnail_url: Vec<String>,
     ///<https://schema.org/copyrightYear>
-    #[serde(default, deserialize_with = "one_or_many")]
+    #[serde(default, deserialize_with = "one_or_many_string_or_num")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub copyright_year: Vec<f32>,
+    pub copyright_year: Vec<i32>,
     ///<https://schema.org/video>
     #[serde(default, deserialize_with = "one_or_many")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
